+++ src/pageModules/LandingModule.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
-import SplitHero from './SplitHero';
-import H1 from './H1';
-import BodyTextLockup from './BodyTextLockup';
+import SplitHero from 'components/SplitHero';
+import H1 from 'components/H1';
+import BodyTextLockup from 'components/BodyTextLockup';
 import mainImage from 'media/bali-landscape-1.jpg';
 
 // mainImage should be managed in JSON
